renamed file and added logic to signIn method
