+++ app/helpers/users/validate_credentials.js
